+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/format/highlimit/HighLimit.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.store.format.busted;
+package org.neo4j.kernel.impl.store.format.highlimit;
 
 import org.neo4j.kernel.impl.store.format.RecordFormat;
 import org.neo4j.kernel.impl.store.format.RecordFormats;
@@ -34,19 +34,19 @@ import org.neo4j.kernel.impl.store.record.RelationshipRecord;
 import org.neo4j.kernel.impl.store.record.RelationshipTypeTokenRecord;
 
 /**
- * Record format with very high limits, 58-bit per ID, while at the same time keeping store size small.
+ * Record format with very high limits, 50-bit per ID, while at the same time keeping store size small.
  *
- * @see BaseBustedRecordFormat
+ * @see BaseHighLimitRecordFormat
  */
-public class Busted implements RecordFormats
+public class HighLimit implements RecordFormats
 {
-    public static final RecordFormats RECORD_FORMATS = new Busted();
+    public static final RecordFormats RECORD_FORMATS = new HighLimit();
 
     @Override
     public String storeVersion()
     {
-        // Community.Busted.Zero
-        return "vC.B.0";
+        // Community.HighLimit.Zero
+        return "vC.H.0";
     }
 
     @Override
