+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/format/highlimit/Reference.java
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.store.format.busted;
+package org.neo4j.kernel.impl.store.format.highlimit;
 
 import java.io.IOException;
