+++ community/kernel/src/main/java/org/neo4j/kernel/impl/store/format/highlimit/SecondaryPageCursorReadDataAdapter.java
@@ -17,13 +17,13 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.kernel.impl.store.format.busted;
+package org.neo4j.kernel.impl.store.format.highlimit;
 
 import java.io.IOException;
 
 import org.neo4j.io.pagecache.PageCursor;
 import org.neo4j.io.pagecache.PagedFile;
-import org.neo4j.kernel.impl.store.format.busted.Reference.DataAdapter;
+import org.neo4j.kernel.impl.store.format.highlimit.Reference.DataAdapter;
 
 /**
  * {@link DataAdapter} able to acquire a secondary {@link PageCursor} on potentially a different page
@@ -59,10 +59,10 @@ class SecondaryPageCursorReadDataAdapter implements DataAdapter<PageCursor>, Sec
             {
                 // Just read out the header, get it out of the way and verify that this secondary record
                 // is in fact a secondary record.
-                // TODO can we do this in BaseBustedRecordFormat (the place where this adapter is created) instead?
+                // TODO can we do this in BaseHighLimitRecordFormat (the place where this adapter is created) instead?
                 byte secondaryHeaderByte = secondaryCursor.getByte();
-                assert (secondaryHeaderByte & BaseBustedRecordFormat.HEADER_BIT_RECORD_UNIT) != 0;
-                assert (secondaryHeaderByte & BaseBustedRecordFormat.HEADER_BIT_FIRST_RECORD_UNIT) == 0;
+                assert (secondaryHeaderByte & BaseHighLimitRecordFormat.HEADER_BIT_RECORD_UNIT) != 0;
+                assert (secondaryHeaderByte & BaseHighLimitRecordFormat.HEADER_BIT_FIRST_RECORD_UNIT) == 0;
                 switched = true;
             }
             return secondaryCursor.getByte();
