+++ community/kernel/src/test/java/org/neo4j/kernel/impl/store/format/HighLimitRecordFormatTest.java
@@ -19,15 +19,15 @@
  */
 package org.neo4j.kernel.impl.store.format;
 
-import org.neo4j.kernel.impl.store.format.busted.Busted;
+import org.neo4j.kernel.impl.store.format.highlimit.HighLimit;
 
-public class BustedRecordFormatTest extends RecordFormatTest
+public class HighLimitRecordFormatTest extends RecordFormatTest
 {
     protected static final RecordGenerators _58_BIT_LIMITS = new LimitedRecordGenerators( random, 58, 58, 58, 16, NULL );
     protected static final RecordGenerators _50_BIT_LIMITS = new LimitedRecordGenerators( random, 50, 50, 50, 16, NULL );
 
-    public BustedRecordFormatTest()
+    public HighLimitRecordFormatTest()
     {
-        super( Busted.RECORD_FORMATS, _50_BIT_LIMITS );
+        super( HighLimit.RECORD_FORMATS, _50_BIT_LIMITS );
     }
 }
